backend/app/matting/service.py
import os
import uuid
from PIL import Image
import time

def process_image(input_path, output_path):
    """
    Process the image to remove background
    
    In a real implementation, this would use a sophisticated algorithm,
    possibly involving machine learning models for image segmentation.
    For this prototype, we'll implement a simple placeholder function.
    
    Args:
        input_path: Path to the original image
        output_path: Path where processed image will be saved
        
    Returns:
        bool: True if processing was successful, False otherwise
    """
    try:
        # Open the input image
        image = Image.open(input_path)
        
        # Simulate processing time
        time.sleep(1)  # Pretend we're doing complex processing
        
        # For this prototype, we'll just create a simple transparency effect
        # In a real implementation, this would be replaced with actual ML-based matting
        
        # Create a transparent version (with white background)
        # Convert to RGBA if it isn't already
        if image.mode != 'RGBA':
            image = image.convert('RGBA')
        
        # Create a new image with transparency
        width, height = image.size
        new_image = Image.new('RGBA', (width, height), (0, 0, 0, 0))
        
        # Process the image - this is a simple placeholder effect
        # A real implementation would use ML for precise segmentation
        pixels = image.load()
        new_pixels = new_image.load()
        
        # Simple processing - make bright pixels more transparent
        # This is just a placeholder, not real background removal
        for y in range(height):
            for x in range(width):
                r, g, b, a = pixels[x, y]
                # Simple adjustment - in a real app, this would be ML-based
                if r > 200 and g > 200 and b > 200:  # If pixel is whitish
                    a = 0  # Make it transparent
                new_pixels[x, y] = (r, g, b, a)
        
        # Save the processed image
        new_image.save(output_path, format='PNG')
        return True
        
    except Exception as e:
        print(f"Error processing image: {e}")
        return False

def generate_unique_filename(filename):
    """Generate a unique filename to avoid conflicts"""
    # Extract file extension
    _, ext = os.path.splitext(filename)
    
    # Generate a unique name using UUID
    unique_name = f"{uuid.uuid4()}{ext}"
    
    return unique_name